refactor into use recommended hook
