+++ src/components/questions/recommended/recommended-questions.tsx
@@ -1,39 +1,12 @@
-import { Question } from "@/types/question";
-import { apiQuestions } from "@/utils/questions-api";
-import { useSession } from "next-auth/react";
-import { useCallback, useEffect, useState } from "react";
+import useRecommended from "@/hooks/use-recommended";
 import QuestionsList from "../questions-list";
 import QuestionsLoading from "../questions-loading";
 import RecommendedHeader from "./recommended-header";
 import ToggleRecommendedButton from "./toggle-recommended-button";
 
 export default function RecommendedQuestions() {
-  const [recommendedQuestions, setRecommendedQuestions] =
-    useState<Question[]>();
-  const [isLoading, setIsLoading] = useState(false);
-  const [isHidden, setIsHidden] = useState(false);
-  const { data: session } = useSession();
-
-  useEffect(() => {
-    (async () => {
-      if (session && session.user && session.user.id) {
-        setIsLoading(true);
-        const data = await apiQuestions.getRecommendedQuestions(
-          session.user.id
-        );
-
-        if (data) {
-          setRecommendedQuestions(data);
-        }
-        setIsLoading(false);
-      }
-    })();
-  }, [session]);
-
-  const toggleIsHidden = useCallback(() => {
-    setIsHidden(!isHidden);
-  }, [isHidden]);
-
+  const { recommendedQuestions, isLoading, isHidden, toggleIsHidden } =
+    useRecommended();
   return (
     <>
       <ToggleRecommendedButton
